requirements/base.txt
docxtpl==0.16.5
Django==4.1.4
django-enumchoicefield=3.0.1
django-environ==0.9.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5